requirements.txt
OdooRPC==0.6.1
packaging==17.1
pluggy==0.7.1
py==1.5.4
pyparsing==2.2.0
python-dotenv==0.9.0
six==1.11.0
tox==3.1.2
virtualenv==16.0.0
